feat: migrate NotFound.spec
